update to fix if compile failed

+++ clang-builder.sh
@@ -11,6 +11,8 @@ git config --global http.sslverify false
 # Set a directory
 DIR="$(pwd ...)"
 EsOne="${1}"
+fail="n"
+TagsDate="$(date +"%Y%m%d")"
 
 if [ "$EsOne" == "13" ];then
     UseBranch="release/13.x"
@@ -54,93 +56,60 @@ fi
     --no-ccache \
     --branch "$UseBranch" \
     --pgo "kernel-defconfig-slim" \
-    "${EXTRA_ARGS[@]}"
+    "${EXTRA_ARGS[@]}" || fail="y"
 
+if [[ "$fail" == "n"]];then
+    # Build binutils
+    ./build-binutils.py --targets aarch64 arm x86_64
 
-# Build binutils
-./build-binutils.py --targets arm aarch64 x86_64
+    # Remove unused products
+    rm -fr install/include
+    rm -f install/lib/*.a install/lib/*.la
 
-# Remove unused products
-rm -fr install/include
-rm -f install/lib/*.a install/lib/*.la
+    # Strip remaining products
+    for f in $(find install -type f -exec file {} \; | grep 'not stripped' | awk '{print $1}'); do
+        strip -s "${f: : -1}"
+    done
 
-# Strip remaining products
-for f in $(find install -type f -exec file {} \; | grep 'not stripped' | awk '{print $1}'); do
-	strip -s "${f: : -1}"
-done
+    # Set executable rpaths so setting LD_LIBRARY_PATH isn't necessary
+    for bin in $(find install -mindepth 2 -maxdepth 3 -type f -exec file {} \; | grep 'ELF .* interpreter' | awk '{print $1}'); do
+        # Remove last character from file output (':')
+        bin="${bin: : -1}"
 
-# Set executable rpaths so setting LD_LIBRARY_PATH isn't necessary
-for bin in $(find install -mindepth 2 -maxdepth 3 -type f -exec file {} \; | grep 'ELF .* interpreter' | awk '{print $1}'); do
-	# Remove last character from file output (':')
-	bin="${bin: : -1}"
+        echo "$bin"
+        patchelf --set-rpath "$DIR/install/lib" "$bin"
+    done
 
-	echo "$bin"
-	patchelf --set-rpath "$DIR/install/lib" "$bin"
-done
+    # Release Info
+    pushd llvm-project || exit
+    llvm_commit="$(git rev-parse HEAD)"
+    short_llvm_commit="$(cut -c-8 <<< "$llvm_commit")"
+    popd || exit
 
-# Release Info
-pushd llvm-project || exit
-llvm_commit="$(git rev-parse HEAD)"
-short_llvm_commit="$(cut -c-8 <<< "$llvm_commit")"
-popd || exit
+    llvm_commit_url="https://github.com/llvm/llvm-project/commit/$short_llvm_commit"
+    binutils_ver="$(ls | grep "^binutils-" | sed "s/binutils-//g")"
+    clang_version="$($DIR/install/bin/clang --version | head -n1 | cut -d' ' -f4)"
+    clang_version_f="$($DIR/install/bin/clang --version | head -n1)"
 
-llvm_commit_url="https://github.com/llvm/llvm-project/commit/$short_llvm_commit"
-binutils_ver="$(ls | grep "^binutils-" | sed "s/binutils-//g")"
-clang_version="$(install/bin/clang --version | head -n1 | cut -d' ' -f4)"
-clang_version_f="$(install/bin/clang --version | head -n1)"
+    git config --global user.name 'ZyCromerZ'
+    git config --global user.email 'neetroid97@gmail.com'
 
-git config --global user.name 'ZyCromerZ'
-git config --global user.email 'neetroid97@gmail.com'
+    ZipName="Clang-$clang_version-${TagsDate}.tar.gz"
+    ClangLink="https://github.com/ZyCromerZ/Clang/releases/download/${clang_version}-${TagsDate}-release/$ZipName"
 
-TagsDate="$(date +"%Y%m%d")"
-ZipName="Clang-$clang_version-${TagsDate}.tar.gz"
-ClangLink="https://github.com/ZyCromerZ/Clang/releases/download/${clang_version}-${TagsDate}-release/$ZipName"
+    pushd $(pwd)/install || exit
+    echo "# Quick Info" > README.md
+    echo "* Build Date : $(date +"%Y-%m-%d")" >> README.md
+    echo "* Clang Version : $clang_version_f" >> README.md
+    echo "* Binutils Version : $binutils_ver" >> README.md
+    echo "* Compiled Based : $llvm_commit_url" >> README.md
+    echo "" >> README.md
+    echo "# link downloads:" >> readme.md
+    echo "* <a href=$ClangLink>$ZipName</a>" >> readme.md
+    tar -czvf ../"$ZipName" *
+    popd || exit
+fi
 
-pushd $(pwd)/install || exit
-echo "# Quick Info" > README.md
-echo "* Build Date : $(date +"%Y-%m-%d")" >> README.md
-echo "* Clang Version : $clang_version_f" >> README.md
-echo "* Binutils Version : $binutils_ver" >> README.md
-echo "* Compiled Based : $llvm_commit_url" >> README.md
-echo "" >> README.md
-echo "# link downloads:" >> readme.md
-echo "* <a href=$ClangLink>$ZipName</a>" >> readme.md
-tar -czvf ../"$ZipName" *
-popd || exit
-
-git clone https://${GIT_SECRET}@github.com/ZyCromerZ/Clang -b main $(pwd)/FromGithub
-pushd $(pwd)/FromGithub || exit
-echo "$(date +"%Y-%m-%d")" > Clang-$EsOne-lastbuild.txt
-echo "$ClangLink" > Clang-$EsOne-link.txt
-git commit -asm "Upload $clang_version_f"
-git checkout -b ${clang_version}-${TagsDate}
-cp ../install/README.md .
-git add .
-git commit -asm "Upload $clang_version_f"
-git tag ${clang_version}-${TagsDate}-release -m "Upload $clang_version_f"
-git push -f origin main ${clang_version}-${TagsDate}
-git push -f origin ${clang_version}-${TagsDate}-release
-popd || exit
-
-chmod +x github-release
-./github-release release \
-    --security-token "$GIT_SECRET" \
-    --user ZyCromerZ \
-    --repo Clang \
-    --tag ${clang_version}-${TagsDate}-release \
-    --name "Clang-${clang_version}-${TagsDate}-release" \
-    --description "$(cat install/README.md)"
-
-fail="n"
-./github-release upload \
-    --security-token "$GIT_SECRET" \
-    --user ZyCromerZ \
-    --repo Clang \
-    --tag ${clang_version}-${TagsDate}-release \
-    --name "$ZipName" \
-    --file "$ZipName" || fail="y"
-
-TotalTry="0"
 UploadAgain()
 {
     fail="n"
@@ -153,28 +122,60 @@ UploadAgain()
         --file "$ZipName" || fail="y"
     TotalTry=$(($TotalTry+1))
     if [ "$fail" == "y" ];then
-        if [ "$TotalTry" != "5" ];then
+        if [ "$TotalTry" != "6" ];then
             sleep 10s
             UploadAgain
         fi
     fi
 }
-if [ "$fail" == "y" ];then
-    sleep 10s
-    UploadAgain
-fi
 
-if [ "$fail" == "y" ];then
+if [[ ! -z "$clang_version" ]];then
+    git clone https://${GIT_SECRET}@github.com/ZyCromerZ/Clang -b main $(pwd)/FromGithub
     pushd $(pwd)/FromGithub || exit
-    git push -d origin ${clang_version}-${TagsDate}
-    git push -d origin ${clang_version}-${TagsDate}-release
-    git checkout main
-    git reset --hard HEAD~1
-    git push -f origin main
+    echo "$(date +"%Y-%m-%d")" > Clang-$EsOne-lastbuild.txt
+    echo "$ClangLink" > Clang-$EsOne-link.txt
+    git commit -asm "Upload $clang_version_f"
+    git checkout -b ${clang_version}-${TagsDate}
+    cp ../install/README.md .
+    git add .
+    git commit -asm "Upload $clang_version_f"
+    git tag ${clang_version}-${TagsDate}-release -m "Upload $clang_version_f"
+    git push -f origin main ${clang_version}-${TagsDate}
+    git push -f origin ${clang_version}-${TagsDate}-release
     popd || exit
-else
-    curl -X POST "https://api.telegram.org/bot$BOT_TOKEN/sendMessage" -d chat_id="-1001628919239" \
-        -d "disable_web_page_preview=true" \
-        -d "parse_mode=html" \
-        -d text="New Toolchain Already Builded boy%0ADate : <code>$(date +"%Y-%m-%d")</code>%0A<code> --- Detail Info About it --- </code>%0AClang version : <code>$clang_version_f</code>%0ABINUTILS version : <code>$binutils_ver</code>%0A%0ALink downloads : <code>$ClangLink</code>%0A%0A-- uWu --"
+
+    chmod +x github-release
+    ./github-release release \
+        --security-token "$GIT_SECRET" \
+        --user ZyCromerZ \
+        --repo Clang \
+        --tag ${clang_version}-${TagsDate}-release \
+        --name "Clang-${clang_version}-${TagsDate}-release" \
+        --description "$(cat install/README.md)"
+
+    # ./github-release upload \
+    #     --security-token "$GIT_SECRET" \
+    #     --user ZyCromerZ \
+    #     --repo Clang \
+    #     --tag ${clang_version}-${TagsDate}-release \
+    #     --name "$ZipName" \
+    #     --file "$ZipName" || fail="y"
+
+    TotalTry="0"
+    UploadAgain
+
+    if [ "$fail" == "y" ];then
+        pushd $(pwd)/FromGithub || exit
+        git push -d origin ${clang_version}-${TagsDate}
+        git push -d origin ${clang_version}-${TagsDate}-release
+        git checkout main
+        git reset --hard HEAD~1
+        git push -f origin main
+        popd || exit
+    else
+        curl -X POST "https://api.telegram.org/bot$BOT_TOKEN/sendMessage" -d chat_id="-1001628919239" \
+            -d "disable_web_page_preview=true" \
+            -d "parse_mode=html" \
+            -d text="New Toolchain Already Builded boy%0ADate : <code>$(date +"%Y-%m-%d")</code>%0A<code> --- Detail Info About it --- </code>%0AClang version : <code>$clang_version_f</code>%0ABINUTILS version : <code>$binutils_ver</code>%0A%0ALink downloads : <code>$ClangLink</code>%0A%0A-- uWu --"
+    fi
 fi
